block: fiops read/write request scale
read/write speed of Flash based storage usually is different. For example,
in my SSD maxium thoughput of read is about 3 times faster than that of
write. Add a scale to differenate read and write. Also add a tunable, so
user can assign different scale for read and write.

By default, the scale is 1:1, which means the scale is a noop.

Change-Id: I7a510a8b2f27d3346972ac402586e6fc984c1747
Signed-off-by: Shaohua Li <shaohua.li@intel.com>